+++ integration-test/integration/chat/github_copilot_test.clj
@@ -123,7 +123,7 @@
           (match-content chat-id "assistant" {:type "reasonStarted" :id (m/pred string?)})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "I should say"})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text " hello"})
-          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?)})
+          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
           (match-content chat-id "assistant" {:type "text" :text "hello"})
           (match-content chat-id "assistant" {:type "text" :text " there!"})
           (match-content chat-id "system" {:type "usage"
@@ -158,7 +158,7 @@
           (match-content chat-id "assistant" {:type "reasonStarted" :id (m/pred string?)})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "I should say"})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text " fine"})
-          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?)})
+          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
           (match-content chat-id "assistant" {:type "text" :text "I'm "})
           (match-content chat-id "assistant" {:type "text" :text " fine"})
           (match-content chat-id "system" {:type "usage"
@@ -207,7 +207,7 @@
           (match-content chat-id "assistant" {:type "reasonStarted" :id (m/pred string?)})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "I should call tool"})
           (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text " eca_directory_tree"})
-          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?)})
+          (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
           (match-content chat-id "assistant" {:type "text" :text "I will list files"})
           (match-content chat-id "assistant" {:type "toolCallPrepare"
                                               :origin "native"
@@ -248,6 +248,7 @@
                                               :id "tool-1"
                                               :name "eca_directory_tree"
                                               :arguments {:path (h/project-path->canon-path "resources")}
+                                              :totalTimeMs number?
                                               :summary "Listing file tree"})
           (match-content chat-id "assistant" {:type "toolCalled"
                                               :origin "native"